BUILD: Add API endpoints and credentials
API endpoints and app secrets are provided through Xcode's build
configuration files for the Debug and Release configurations. The values
are then passed to the app at build time through the Info.plist file.

The configuration files are not committed to the repository to avoid
compromising the app secrets. Developers working on the project will
need to recreate these files in their working copy or get them from a
shared secure storage.